Correct use of Unwrap return value to determine success
Previously, SLIPUnpacketize() and other "unpacketize" functions would convert the return value from Unwrap directly to a boolean. This is incorrect, as Unwrap returns a negative value if there is an error, otherwise it returns the data size. So it would fail on valid packets with no data, but report success on actual errors. The correct way to interpret whether a call to Unwrap was successful, is to return `Unwrap() >= 0`.